op-deployer,ctb: Migrate DeployPreimageOracle to new script API
